Backend/requirements.txt
fastapi
uvicorn
faiss-cpu
numpy
pandas
groq
python-dotenv
mysql-connector-python
bcrypt
PyJWT
